cleanup a bit, regenerate example code, add more todos
