+++ floor_generator/lib/processor/query_method_processor.dart
@@ -6,8 +6,8 @@ import 'package:floor_generator/misc/constants.dart';
 import 'package:floor_generator/misc/type_utils.dart';
 import 'package:floor_generator/processor/error/query_method_processor_error.dart';
 import 'package:floor_generator/processor/processor.dart';
+import 'package:floor_generator/processor/query_analyzer/analyzed_query.dart';
 import 'package:floor_generator/processor/query_analyzer/engine.dart';
-import 'package:floor_generator/processor/query_analyzer/sqlite_types.dart';
 import 'package:floor_generator/value_object/query_method.dart';
 import 'package:floor_generator/value_object/query_method_return_type.dart';
 import 'package:floor_generator/value_object/queryable.dart';
